+++ arduino/builder/linker.go
@@ -13,7 +13,7 @@
 // Arduino software without disclosing the source code of your own applications.
 // To purchase a commercial license, send an email to license@arduino.cc.
 
-package phases
+package builder
 
 import (
 	"bytes"
